requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
